Standardized build and updated README.md
Updated the rakefile to use the standard build process for Mvba and added
necessary supporting files and removed files used by old build process
(build_support/).

Cleaned up the README.md and added section on how to build the project.
Updated links in the license section to refer to GitHub instead of Google
Code.

Updated the supporting dlls:

Updated FastReflection to 1.0.0.13
Updated FluentAssert to 1.0.0.3
Updated LinqToHtml to 1.0.0.1
Updated MvbaCore to 1.0.0.1
Updated StructureMap to 2.6.1.0